Add files via upload
If you know how to run Python code this will make it VERY easy for you to set up a 256-bit encryption capability on your computer. 

The program asks you only two questions: it basically asks you to input a sentence, or a string of any ascii characters, or whatever you wish, that is at least 32 characters long. It then asks you to write the message that you want encoded.  It then spits out the encrypted message. That's all it takes